wip: move old/new impl into separate folders
