+++ src/LEGACY_IMPL/LEGACY_PagerViewNativeComponent.ts
