+++ src/NEW_IMPL/PagerViewNativeComponent.ts
